requirements.txt
darts==0.23.0
eco2ai==0.3.9
filterpy==1.4.5
hydra-core==1.3.2
matplotlib==3.7.1
numpy==1.23.5
omegaconf==2.3.0
optuna==3.1.1
pandas==1.4.3
pytorch_lightning==1.7.7
scipy==1.9.1
seaborn==0.12.2
torch==2.0.0
torchaudio==2.0.1
tqdm==4.65.0
